requirements.txt
Flask
feedparser
python-dotenv
ollama
websocket-client
